Update to develop for cleanup and use of entry.OnSubmitted
