+++ go.mod
@@ -3,7 +3,7 @@ module github.com/Jacalz/wormhole-gui
 go 1.14
 
 require (
-	fyne.io/fyne v1.4.4-0.20210109154404-fc141cc2c0c5
+	fyne.io/fyne v1.4.4-0.20210112162126-680cb739a946
 	github.com/mholt/archiver/v3 v3.5.0
 	github.com/psanford/wormhole-william v1.0.5-0.20200920232445-f69f6e823d8c
 )
